rename rr -> rules
