+++ setup.py
@@ -83,7 +83,7 @@ modules = [
     'sympy.polys.domains',
     'sympy.printing',
     'sympy.printing.pretty',
-    'sympy.rr',
+    'sympy.rules',
     'sympy.series',
     'sympy.sets',
     'sympy.simplify',
