+++ sympy/rules/strat_pure.py
