+++ sympy/rules/traverse.py
